add some fixes
