+++ homework4/tests/HW4_test_task3.py
@@ -2,14 +2,12 @@ from homework4.task3.get_print_output import my_precious_logger
 
 
 class TestStderrStdout:
-    def test_stderr_log(self, capsys):
-        """Testing that stderr equal 'error' if string start with that"""
-        my_precious_logger("error")
+    def test_error_if_string_start_with_error(self, capsys):
+        my_precious_logger("error of all time")
         out, err = capsys.readouterr()
-        assert err == "error"
+        assert err == 'error: file not found'
 
-    def test_stdout_log(self, capsys):
-        """Testing that stdout equal string"""
-        my_precious_logger("Ok")
+    def test_text_not_start_with_error(self, capsys):
+        my_precious_logger("it fine, im ok")
         out, err = capsys.readouterr()
-        assert out == "Ok"
+        assert out == "it fine, im ok"
